requirements.txt
django===4.0.6
djangorestframework==3.14.0
psycopg2-binary==2.9.5
